% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{basic_heatmap}
\alias{basic_heatmap}
\title{Heatmap of the data}
\usage{
basic_heatmap(data, heatmap_color_select)
}
\arguments{
\item{data}{A basic data matrix}

\item{heatmap_color_select}{Vector of colors to use for the fill
in the heatmap}
}
\value{
A drawn ComplexHeatmap.
}
\description{
Create a ComplexHeatmap from a data matrix.
}
