% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_08_bicluster.R
\name{bicluster_summary_message}
\alias{bicluster_summary_message}
\title{Summary of biclust}
\usage{
bicluster_summary_message(biclustering, select_bicluster)
}
\arguments{
\item{biclustering}{Return list from \code{get_biclustering} function}

\item{select_bicluster}{Selected cluster to return information on}
}
\value{
A character string message with the requested summary information.
}
\description{
Give the summary of the identified clusters from the \code{get_biclustering}
function which gives the number of clusters found and the samples used.
}
