% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{change_names}
\alias{change_names}
\title{Change comparison names}
\usage{
change_names(comparison)
}
\arguments{
\item{comparison}{Comparison to change the name for}
}
\value{
The altered comparison string, see the description for example.
}
\description{
Change comparison names in limma from "KO_ko-WT_ko" to    "KO-WT_for_ko"
}
