% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_07_genome.R
\name{chromosome_plotly}
\alias{chromosome_plotly}
\title{Plotly of chromosome position}
\usage{
chromosome_plotly(
  limma,
  select_contrast,
  all_gene_info,
  ignore_non_coding,
  limma_p_val_viz,
  limma_fc_viz,
  label_gene_symbol,
  ma_window_size,
  ma_window_steps,
  ch_region_p_val
)
}
\arguments{
\item{limma}{Return from \code{limma_value} function}

\item{select_contrast}{DEG contrast to examine}

\item{all_gene_info}{Gene information return from \code{get_gene_info}}

\item{ignore_non_coding}{When TRUE only use protein coding genes}

\item{limma_p_val_viz}{Adjusted p-value to use for significant genes}

\item{limma_fc_viz}{Minimum fold-change value to filter with}

\item{label_gene_symbol}{Paste the gene symbol label on the plot}

\item{ma_window_size}{Moving average window size for a chromosome 
(1, 2, 4, 6, 8, 10, 15, 20)}

\item{ma_window_steps}{Number of moving average window steps (1, 2, 3, 4)}

\item{ch_region_p_val}{P-value to use for finding significant chromosome
region enrichment}
}
\value{
Plotly visualization of chromosomes and significantly enriched genes
}
\description{
Calculate a plotly that shows the chromosome position of significant genes
and significantly enriched regions.
}
