% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{cluster_heat_click_info}
\alias{cluster_heat_click_info}
\title{Interactive click text for subheatmap}
\usage{
cluster_heat_click_info(
  click,
  ht_sub,
  ht_sub_obj,
  ht_pos_sub,
  sub_groups,
  group_colors,
  cluster_meth,
  click_data
)
}
\arguments{
\item{click}{Click input from subheatmap}

\item{ht_sub}{Drawn subheatmap}

\item{ht_sub_obj}{Heatmap object with mapping info}

\item{ht_pos_sub}{Position information from submap}

\item{sub_groups}{Vector of group labels from submap}

\item{group_colors}{Colors for the group annotation}

\item{cluster_meth}{Type of clustering being performed}

\item{click_data}{Data matrix to get the data value from}
}
\value{
HTML code to produce a table with information
 about the selected cell.
}
\description{
Create a text output to tell the user the cell
information for their click.
}
