% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{colorpanel2}
\alias{colorpanel2}
\title{PATHVIEW FUNCTION
https://rdrr.io/bioc/pathview/src/R/colorpanel2.R}
\usage{
colorpanel2(n, low, mid, high)
}
\description{
PATHVIEW FUNCTION
https://rdrr.io/bioc/pathview/src/R/colorpanel2.R
}
