% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_01_load_data.R
\name{convert_data}
\alias{convert_data}
\title{Convert expression data rownames to ensembl}
\usage{
convert_data(converted, data, no_id_conversion)
}
\arguments{
\item{converted}{Data from convert_id function containing converted ids}

\item{data}{Data from inputed expression file}

\item{no_id_conversion}{TRUE/FALSE for converting data ids or not}
}
\value{
Returns original data with rownames converted to ensembl
}
\description{
This function works with the converted ids to take in the
expression data and swap the rownames to the converted ids.
It returns the data exactly the same except for the changed
ids.
}
