% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{cor_plot}
\alias{cor_plot}
\title{Create a correlation plot from heatmap data}
\usage{
cor_plot(data, label_pcc, heat_cols, text_col)
}
\arguments{
\item{data}{Heatmap data}

\item{label_pcc}{Label with correlation coefficient when TRUE}

\item{heat_cols}{Heat colors to use in the correlation matrix}

\item{text_col}{Color to make the text labels in the plot}
}
\value{
ggplot heatmap of correlation matrix
}
\description{
Creates a correlation matrix heatmap from the
heatmap data to demonstrate the correlation
between samples.
}
