% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{deg_heat_data}
\alias{deg_heat_data}
\title{Find data for the heatmap}
\usage{
deg_heat_data(limma, select_contrast, processed_data, contrast_samples)
}
\arguments{
\item{limma}{Return results list from the limma_value function}

\item{select_contrast}{Comparison from DEG analysis to filter
for the significant genes}

\item{processed_data}{Data that has been through the pre-processing}

\item{contrast_samples}{Columns that are in the group of the
selected comparison}
}
\value{
Submatrix of the processed data with only the significantly
 expressed genes and the columns that are in the selected contrast
 group.
}
\description{
Filter the processed data into a submatrix that only contains
the genes that had a significant fold change. The samples will
also be subsetted to only contain samples that pertain to the
selected comparison or contrast.
}
