% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{deg_heat_sub}
\alias{deg_heat_sub}
\title{Plot brush selection from main heatmap}
\usage{
deg_heat_sub(ht_brush, ht, ht_pos_main, heatmap_data, all_gene_names)
}
\arguments{
\item{ht_brush}{Input from the user creating a brush selection
on the main heatmap}

\item{ht}{Main heatmap from the deg_heatmap function}

\item{ht_pos_main}{Main heatmap position information to use
for the sub heatmap}

\item{heatmap_data}{Original data matrix that was plotted in
the main heatmap}

\item{all_gene_names}{Data matrix of all the mapped gene names}
}
\value{
A ComplexHeatmap object of the brushed selection from
 the main heatmap.
}
\description{
Create a ComplexHeatmap object from the User brush selection
that is a sub plot of the main plot.
}
