% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{deg_heatmap}
\alias{deg_heatmap}
\title{Heatmap for significant contrast genes}
\usage{
deg_heatmap(data, bar, heatmap_color_select)
}
\arguments{
\item{data}{Submatrix of the processed data matrix from the
deg_heta_data function}

\item{bar}{Vector to signify a positive (1) expression fold
change or a negative (-1) change}

\item{heatmap_color_select}{Color vector to use for the
heatmap expression scale}
}
\value{
A drawn heatmap from the filtered data.
}
\description{
Create a ComplexHeatmap of the processed expression data for
the genes that were significantly expressed in the selected 
comparison. The data for this heatmap comes from the
deg_heat_data function.
}
