% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{deg_limma}
\alias{deg_limma}
\title{Differential expression using limma package}
\usage{
deg_limma(
  processed_data,
  max_p_limma = 0.1,
  min_fc_limma = 2,
  raw_counts,
  counts_deg_method,
  prior_counts,
  data_file_format,
  selected_comparisons = NULL,
  sample_info = NULL,
  model_factors = NULL,
  block_factor = NULL
)
}
\arguments{
\item{processed_data}{Data that has been through the pre-processing}

\item{max_p_limma}{Significant p-value to use for the fold-change
values}

\item{min_fc_limma}{Minimum fold-change to include in the results}

\item{raw_counts}{The matrix of counts before processing for
gene expression data}

\item{counts_deg_method}{The method or package being used for
the DEG analysis}

\item{prior_counts}{The constant added to the log transformation
from pre-processing}

\item{data_file_format}{Type of gene data being examined}

\item{selected_comparisons}{Selected comparisons to analyze
in the DEG analysis}

\item{sample_info}{Experiment file information for grouping}

\item{model_factors}{Vector of selected factors and interaction terms
from the model design}

\item{block_factor}{The selected factors for batch effect}
}
\value{
The return value is the results of the DEG analysis. These
 results are filtered and formatted by the limma_value function.
}
\description{
Used in the limma_value function to perform DEG analysis using the 
limma package. It is not recommended to use this function on its own.
}
