% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{detect_groups}
\alias{detect_groups}
\title{FUNCTION_TITLE}
\usage{
detect_groups(sample_names, sample_info = NULL)
}
\arguments{
\item{sample_names}{DESCRIPTION.}

\item{sample_info}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
