% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_10_doc.R
\name{doc_hist}
\alias{doc_hist}
\title{Density plot for the processed data}
\usage{
doc_hist(processed_data, sampleID)
}
\arguments{
\item{processed_data}{Data that has gone through the pre-processing}

\item{sample_info}{Sample_info from the experiment file}
}
\value{
Returns a formatted gg density plot
}
\description{
This function takes in the processed data and sample info
and creates a density plot for the distribution of sequences
that are mapped to each sample.
}
