% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{draw_sample_tree}
\alias{draw_sample_tree}
\title{Draw a dendogram of data samples}
\usage{
draw_sample_tree(
  tree_data,
  gene_centering,
  gene_normalize,
  sample_centering,
  sample_normalize,
  hclust_funs,
  hclust_function,
  dist_funs,
  dist_function
)
}
\arguments{
\item{tree_data}{Data that has been through pre-processing}

\item{gene_centering}{TRUE/FALSE subtract mean from gene rows}

\item{gene_normalize}{TRUE/FALSE divide by SD in gene rows}

\item{sample_centering}{TRUE/FALSE subtract mean from sample columns}

\item{sample_normalize}{TRUE/FALSE divide by SD in sample columns}

\item{hclust_funs}{Clustering functions defined in idep}

\item{hclust_function}{String of chosen clustering method}

\item{dist_funs}{Distance functions defined in idep}

\item{dist_function}{Selected distance function}
}
\value{
Dendogram plot of dataset samples
}
\description{
Create a clustered tree of the samples in the dataset.
}
