% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{eda_density}
\alias{eda_density}
\title{Density plot for the processed data}
\usage{
eda_density(processed_data, sample_info)
}
\arguments{
\item{processed_data}{Data that has gone through the pre-processing}

\item{sample_info}{Sample_info from the experiment file}
}
\value{
Returns a formatted gg density plot
}
\description{
This function takes in the processed data and sample info
and creates a density plot for the distribution of sequences
that are mapped to each sample.
}
