% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{eda_scatter}
\alias{eda_scatter}
\title{Scatterplot for EDA on processed data}
\usage{
eda_scatter(processed_data, plot_xaxis, plot_yaxis)
}
\arguments{
\item{processed_data}{Data that has gone through the pre-processing}

\item{plot_xaxis}{Sample to plot on the x-axis}

\item{plot_yaxis}{Sample to plot on the y axis}
}
\value{
Returns a formatted gg scatterplot
}
\description{
This function takes the data after it has been pre-processed and
creates a scatterplot of the counts for two samples that are
selected by the user.
}
