% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{enrich_net}
\alias{enrich_net}
\title{numChar=100 maximum number of characters
n=200  maximum number of nodes
degree.cutoff = 0    Remove node if less connected
from PPInfer}
\usage{
enrich_net(
  data,
  gene_set,
  node_id,
  node_name = node_id,
  p_value,
  n = 50,
  num_char = NULL,
  p_value_cutoff = 0.05,
  edge_cutoff = 0.05,
  degree_cutoff = 0,
  edge_width = function(x) {     5 * x^2 },
  node_size = function(x) {     2.5 * log10(x) },
  group = FALSE,
  group_color = c("green", "red"),
  group_shape = c("circle", "square"),
  legend_parameter = list("topright"),
  show_legend = TRUE,
  plotting = TRUE,
  layout_button = 0,
  ...
)
}
\description{
numChar=100 maximum number of characters
n=200  maximum number of nodes
degree.cutoff = 0    Remove node if less connected
from PPInfer
}
