% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{extract_word}
\alias{extract_word}
\title{FUNCTION_TITLE}
\usage{
extract_word(word_list)
}
\arguments{
\item{word_list}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
