% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{fct_02_pre_process.R}
\alias{fct_02_pre_process.R}
\title{fct_02_pre_process.R This file holds all of the main data analysis functions
associated with second tab of the iDEP website.}
\description{
fct_02_pre_process.R This file holds all of the main data analysis functions
associated with second tab of the iDEP website.
}
\section{fct_02_pre_process.R functions}{

\code{plot_genes}
}

