% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{fct_03_heatmap.R}
\alias{fct_03_heatmap.R}
\title{fct_03_heatmap.R This file holds all of the main data analysis functions
associated with third tab of the iDEP website.}
\description{
fct_03_heatmap.R This file holds all of the main data analysis functions
associated with third tab of the iDEP website.
}
\section{fct_03_heatmap.R functions}{

}

