% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{fct_05_pca.R}
\alias{fct_05_pca.R}
\title{fct_05_pca.R This file holds all of the main data analysis functions
associated with fifth tab of the iDEP website.}
\description{
fct_05_pca.R This file holds all of the main data analysis functions
associated with fifth tab of the iDEP website.
}
\section{fct_05_pca.R functions}{

\code{my_pgsea}
}

