% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{fct_06_deg1.R}
\alias{fct_06_deg1.R}
\title{fct_06_deg1.R This file holds all of the main data analysis functions
associated with sixth tab of the iDEP website.}
\description{
fct_06_deg1.R This file holds all of the main data analysis functions
associated with sixth tab of the iDEP website.
}
\section{fct_06_deg1.R functions}{

\code{change_names}
}

