% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_07_genome.R
\name{fct_07_genome.R}
\alias{fct_07_genome.R}
\title{fct_07_genome.R This file holds all of the main data analysis functions
associated with eighth tab of the iDEP website.}
\description{
fct_07_genome.R This file holds all of the main data analysis functions
associated with eighth tab of the iDEP website.
}
\section{fct_07_genome.R functions}{

}

