% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_08_bicluster.R, R/fct_09_network.R
\name{fct_08_bicluster.R}
\alias{fct_08_bicluster.R}
\title{fct_08_bicluster.R This file holds all of the main data analysis functions
associated with eighth tab of the iDEP website.}
\description{
fct_08_bicluster.R This file holds all of the main data analysis functions
associated with eighth tab of the iDEP website.

fct_08_bicluster.R This file holds all of the main data analysis functions
associated with eighth tab of the iDEP website.
}
\section{fct_07_bicluster.R functions}{



}

