% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{fct_database.R}
\alias{fct_database.R}
\title{fct_database.R: A fie with all functions to deal with database}
\description{
fct_database.R: A fie with all functions to deal with database
}
\section{fct_database.R functions}{

\code{connect_convert_db} connects to the converID.db
and returns the objects.
}

