% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis_random.R
\name{find_contrast_samples}
\alias{find_contrast_samples}
\title{Determine samples in selected contrast}
\usage{
find_contrast_samples(
  select_contrast,
  all_sample_names,
  sample_info = NULL,
  select_factors_model = NULL,
  select_model_comprions = NULL,
  reference_levels = NULL,
  counts_deg_method = NULL,
  data_file_format = NULL
)
}
\arguments{
\item{select_contrast}{Comparison from DEG analysis to filter
for the significant genes}

\item{all_sample_names}{The column names of the processed data}

\item{sample_info}{Experiment file information for grouping
samples}

\item{select_factors_model}{The selected factors for the model
expression}

\item{select_model_comprions}{Selected comparisons to analyze
in the DEG analysis}

\item{reference_levels}{Vector of reference levels to use for the
selected factors}

\item{counts_deg_method}{Method of DEG being performed (See
DEG UI for options)}

\item{data_file_format}{Type of data being examined}
}
\value{
A numeric vector that can be used to index the processed
 data and subset to only include the columns from the selected
 contrast.
}
\description{
Find the samples that are in the group of the selected
contrast. Can be used to subset the data to only include
the samples that correspond to the chosen comparison.
}
