% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{find_id_type_by_id}
\alias{find_id_type_by_id}
\title{Find the ID type}
\usage{
find_id_type_by_id(id, id_index)
}
\arguments{
\item{id}{ID to find the type from}

\item{id_index}{Index of IDs to use}
}
\description{
Using an inputted ID, find the type of IDs
}
