% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis_random.R
\name{find_overlap_gmt}
\alias{find_overlap_gmt}
\title{Calculate overlap for species with GMT file}
\usage{
find_overlap_gmt(
  query,
  gene_set,
  min_fdr = 0.2,
  min_size = 2,
  max_size = 10000
)
}
\arguments{
\item{query}{Vector of genes to calculate the overlap on}

\item{gene_set}{Gene sets from the custom GMT file to use in
the overlap}

\item{min_fdr}{Significant p-value to determine significantly
expressed genes}

\item{min_size}{Minimum size for a pathway gene set}

\item{max_size}{Maximum size for a pathway gene set}
}
\description{
For a species not in the iDEP database, calculate the overlap
 for pathway analysis.
}
