% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{find_species_by_id}
\alias{find_species_by_id}
\title{Find a species by ID}
\usage{
find_species_by_id(species_id, org_info)
}
\arguments{
\item{species_id}{Species ID to search the database with}

\item{org_info}{iDEP data org_info file}
}
\value{
Species information in \code{org_info} from the
 matched ID.
}
\description{
Find a species in the iDEP database with an
ID.
}
