% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{find_species_by_id_name}
\alias{find_species_by_id_name}
\title{Find a species by ID}
\usage{
find_species_by_id_name(species_id, org_info)
}
\arguments{
\item{species_id}{Species ID to search the database with}

\item{org_info}{iDEP data org_info file}
}
\value{
Only return the species name with this function.
}
\description{
Find a species in the iDEP database with an
ID.
}
