% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{gage_data}
\alias{gage_data}
\title{Pathway analysis with gage package}
\usage{
gage_data(
  select_go,
  select_contrast,
  min_set_size,
  max_set_size,
  limma,
  gene_p_val_cutoff,
  gene_sets,
  absolute_fold,
  pathway_p_val_cutoff,
  n_pathway_show
)
}
\arguments{
\item{select_go}{The portion of the database to use for
pathway analysis}

\item{select_contrast}{Comparison from DEG analysis to
use the top genes from in the pathway analysis}

\item{min_set_size}{Minimum gene set size for a pathway}

\item{max_set_size}{Maximum gene set size for a pathway}

\item{limma}{Results list from the limma_value function}

\item{gene_p_val_cutoff}{Significant p-value to filter
the top genes fold change by}

\item{gene_sets}{List of vectors with each vector being the
set of genes that correspond to a particular pathway in
the database (from read_gene_sets function)}

\item{absolute_fold}{Use the absolute value of the fold
change}

\item{pathway_p_val_cutoff}{Significant p-value to determine
enriched pathways}

\item{n_pathway_show}{Number of pathways to return in final
result}
}
\value{
A data frame with the results of the pathway analysis.
 The data frame has five columns for the direction of the
 regulation, the pathway description, the stat value, the
 number of overlapping genes, and the p-value.
}
\description{
Run pathway analysis with the gage package using the results
from the limma_value function.
}
