% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{genes_stat_table}
\alias{genes_stat_table}
\title{Create a table from DEG results}
\usage{
genes_stat_table(limma)
}
\arguments{
\item{limma}{Return list from the limma_value function}
}
\description{
Using the limma_value return list, create a table of the
number of significantly expressed genes for each analyzed
comparison.
}
