% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_01_load_data.R
\name{get_all_gene_names}
\alias{get_all_gene_names}
\title{Get all matched gene names}
\usage{
get_all_gene_names(mapped_ids, all_gene_info)
}
\arguments{
\item{mapped_ids}{Matched IDs from the convert_data functions}

\item{all_gene_info}{Gene information matched species in idep database}
}
\value{
Data frame containing all the matched ID names from idep
database. Three columns denotes a recognized species for which
idep had gene names for. Two columns means the IDs were converted
to ensembl format, but no species was found for the gene names.
One means no conversion occurred.
}
\description{
This function will create a data frame with all the matched
gene names from the database.
}
