% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_07_genome.R
\name{get_genome_plot}
\alias{get_genome_plot}
\title{PREDA PACKAGE ERRORR #########################}
\usage{
get_genome_plot(genome_plot_data, regions_p_val_cutoff, statistic_cutoff)
}
\description{
PREDA PACKAGE ERRORR #########################
}
