% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_07_genome.R
\name{get_genome_plot_data}
\alias{get_genome_plot_data}
\title{PREDA PACKAGE ERROR ###########################}
\usage{
get_genome_plot_data(
  genome_plot_data_pre,
  all_gene_info,
  select_contrast,
  limma,
  regions_p_val_cutoff,
  statistic_cutoff
)
}
\description{
PREDA PACKAGE ERROR ###########################
}
