% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_07_genome.R
\name{get_genome_plot_data_pre}
\alias{get_genome_plot_data_pre}
\title{PREDA PACKAGE ERROR ########################}
\usage{
get_genome_plot_data_pre(select_contrast, limma, all_gene_info)
}
\description{
PREDA PACKAGE ERROR ########################
}
