% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_09_network.R
\name{get_network_plot}
\alias{get_network_plot}
\title{Network of top genes}
\usage{
get_network_plot(
  select_wgcna_module,
  wgcna,
  top_genes_network,
  select_go,
  select_org,
  all_gene_info,
  edge_threshold
)
}
\arguments{
\item{select_wgcna_module}{The module to create a plot of hte top genes for,
options can be found with the \code{get_wgcna_modules} function}

\item{wgcna}{List returned from the \code{get_wgcna}}

\item{top_genes_network}{Number of genes to include in the network plot}

\item{select_go}{Portion of the database to use in enrichment querying}

\item{select_org}{Organism the expression data is for}

\item{all_gene_info}{Gene info that was found from querying the database}

\item{edge_threshold}{Wavlue from 1-.1 (.4 recommended)}
}
\value{
A function that can be stored as an object and then called to produce
 the plot that the function created. If it is note stored and called the
 function will only return another funciton.
}
\description{
Create a network plot of the top genes found with the WGCNA package.
}
