% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_09_network.R
\name{get_wgcna}
\alias{get_wgcna}
\title{Utilize WGCNA function}
\usage{
get_wgcna(data, n_genes, soft_power, min_module_size)
}
\arguments{
\item{data}{Data matrix that has been through the pre-processing}

\item{n_genes}{Number of most variable genes to use in the function}

\item{soft_power}{Value between 1-20 
(https://support.bioconductor.org/p/87024/)}

\item{min_module_size}{For modules detected by WGCNA, set a minimum size}
}
\value{
A list of 8 objects. \code{data} is a submatrix of the input
 parameter \code{data} which on contains the genes that were selected with
 \code{n_genes}. \code{powers} is a numeric vector from 1-10 and then the
 even numbers from 10-20. \code{sft} is a return from the WGCNA package that
 is a list containing \code{powerEstimate} and \code{fitIndices}. For
 information on these objects visit 
 https://www.rdocumentation.org/packages/WGCNA/versions/1.70-3/topics/pickSoftThreshold.
 \code{tom} is another return from the WGCNA package, for information on this
 matrix visit https://www.rdocumentation.org/packages/WGCNA/versions/1.70-3/topics/TOMsimilarityFromExpr.
 \code{dynamic_colors} is a vector of colors created by the WGCNA package to
 correspond to the modules that were identified by the pacakage.
 \code{module_info} is a data frame with a gene name, a color, and the
 identified module for the gene. \code{n_modules} gives the number of modules
 calculated. \code{n_genes} tells the number of genes included in the
 modules.
}
\description{
Run the WGCNA package on the processed data.
}
