% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_09_network.R
\name{get_wgcna_modules}
\alias{get_wgcna_modules}
\title{List WGCNA modules}
\usage{
get_wgcna_modules(wgcna)
}
\arguments{
\item{wgcna}{List returned from the \code{get_wgcna}}
}
\value{
A character vector with all the strings that can be filled into the
 inut parameter \code{select_wgcna_module} in other WGCNA functions.
}
\description{
Get the options for modules to select from running the \code{get_wgcna}
function.
}
