% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{ggplot colors function}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{Number of colors to return}
}
\value{
Vector of hex color codes for a plot.
}
\description{
This function will return the colors that
the ggplot2 package uses for plots.
}
