% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{heat_sub}
\alias{heat_sub}
\title{Draw sub heatmap from brush input}
\usage{
heat_sub(
  ht_brush,
  ht,
  ht_pos_main,
  heatmap_data,
  sample_info,
  select_factors_heatmap,
  cluster_meth
)
}
\arguments{
\item{ht_brush}{Brush input from the main heatmap}

\item{ht}{Main heatmap object}

\item{ht_pos_main}{Position of brush on main heatmap}

\item{heatmap_data}{Data for the heatmap}

\item{sample_info}{Experiment design file}

\item{select_factors_heatmap}{Group design to label by}

\item{cluster_meth}{Type of clustering being performed}
}
\value{
A list containing a Heatmap from the brush selection
 of the main heatmap, the submap data matrix, the groups for
 the submap, the submap legend, and data for the click info.
}
\description{
Use the brush input from the main heatmap to
create a larger subheatmap.
}
