% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{individual_plots}
\alias{individual_plots}
\title{Individual plotting function for genes}
\usage{
individual_plots(
  individual_data,
  sample_info,
  selected_gene,
  gene_plot_box,
  use_sd,
  lab_rotate
)
}
\arguments{
\item{sample_info}{Experiment file information for grouping samples}

\item{gene_plot_box}{T/F for individual sample plot or grouped data plot}

\item{use_sd}{T/F for standard error or standard deviation bars on bar plot}

\item{merged_data}{Data that has been merged with the gene info}

\item{select_dene}{Gene(s) to be plotted}

\item{select_org}{Species the expression data is from}
}
\value{
A formatted ggplot. For gene_plot_box = TRUE the return will be a
 lineplot for the expression of each individual sample for the selected gene.
 If gene_plot_box = FALSE the return will be a barplot for the groups provided
 in the sample information.
}
\description{
Takes in the merged data and other data to provide
plots on individual gene names. Depening on the selections
this function will either return a barplot from the
grouped data or a lineplot from the individual sample.
}
