% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{k_means_elbow}
\alias{k_means_elbow}
\title{Draw an elbow plot for k-cluster selection}
\usage{
k_means_elbow(heatmap_data)
}
\arguments{
\item{heatmap_data}{Processed heatmap data}
}
\value{
Formatted elbow plot for the data
}
\description{
This function takes in the processed heatmap data and
creates an elbow plot to guide the selection of the
number of clusters to create
}
