% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{kegg_pathway}
\alias{kegg_pathway}
\title{Use KEGG to create a pathway diagram}
\usage{
kegg_pathway(
  go,
  gage_pathway_data,
  sig_pathways,
  select_contrast,
  limma,
  converted,
  idep_data,
  select_org,
  low_color = "green",
  high_color = "red"
)
}
\arguments{
\item{go}{Portion of the database to use for the pathway analysis}

\item{gage_pathway_data}{Return matrix from \code{gage_data}
function}

\item{sig_pathways}{Description of the pathway for which to
obtain the gene expression data}

\item{select_contrast}{Comparison from DEG analysis to
use the top genes from in the pathway analysis}

\item{limma}{Results list from the \code{limma_value}
function}

\item{converted}{Return value from the convert_id function, contains
information about the gene IDs for the matched species}

\item{idep_data}{Read data files from the database}

\item{select_org}{The organism that the gene data is for}

\item{low_color}{Color value for the low-ly expressed genes}

\item{high_color}{Color vlaue for the high-ly expressed genes}
}
\value{
Make an image and return the path to the image to be
 rendered in the server.
}
\description{
In the databse, use the KEGG information to create an image
that is a diagram of the pathway that is being enriched.
}
