% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{kegg_pathway_id}
\alias{kegg_pathway_id}
\title{Find pathway IDs for a KEGG description}
\usage{
kegg_pathway_id(
  pathway_description,
  Species,
  GO,
  select_org,
  gmt_files,
  org_info,
  idep_data
)
}
\arguments{
\item{pathway_description}{Description for a pathway in the
iDEP database}

\item{Species}{Species that the pathway is for}

\item{GO}{Portion of the database to query ("KEGG")}

\item{select_org}{The organism that the gene data is for}

\item{gmt_files}{GMT files from the iDEP database}

\item{org_info}{Organism information files from the iDEP
database}

\item{idep_data}{Background idep_data}
}
\value{
Return the KEGG ID for the pathway.
}
\description{
From a pathway description find the KEGG ID.
}
