% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{limma_value}
\alias{limma_value}
\title{DEG analysis function}
\usage{
limma_value(
  data_file_format,
  counts_deg_method,
  raw_counts,
  limma_p_val,
  limma_fc,
  select_model_comprions,
  sample_info,
  select_factors_model,
  select_interactions,
  select_block_factors_model,
  factor_reference_levels,
  processed_data,
  counts_log_start,
  p_vals
)
}
\arguments{
\item{data_file_format}{Type of gene data being examined}

\item{counts_deg_method}{The method or package being used for
the DEG analysis}

\item{raw_counts}{The matrix of counts before processing for
gene expression data}

\item{limma_p_val}{Significant p-value to use for expressed
genes}

\item{limma_fc}{Minimum fold-change cutoff for the DEG
analysis}

\item{select_model_comprions}{Selected comparisons to analyze
in the DEG analysis}

\item{sample_info}{Experiment file information for grouping}

\item{select_factors_model}{The selected factors for the model
expression}

\item{select_interactions}{The interaction terms being used in
the model design}

\item{select_block_factors_model}{The selected factors for
batch effect}

\item{factor_reference_levels}{Vector of reference levels to
use for the selected factors}

\item{processed_data}{Data that has been through the pre-processing}

\item{counts_log_start}{The constant added to the log transformation
from pre-processing}

\item{p_vals}{The vector of p-vals calculated in pre-process for
significant expression}
}
\value{
List with the results of the DEG analysis. When the function
 is successful there are four entries in the list. "results" is a
 matrix with the same dimensions as the processed data. The entries
 in "results" are c(-1, 0, 1) for (negative fold change, no significant
 change, positive fold change) respectively. The second entry is
 "comparisons" and is a character vector of the different comparisons
 that were analyzed in the function. Third is "exp_type" and details
 the model expression that was used for the DEG analysis. Lastly is
 "top_genes" which is itself a list. The "top_genes" list has an entry
 for each comparison. Each entry is a data frame with two columns. One
 column is the calculated fold change for the comparison and the other
 is the adjusted p-value for the fold change calculation.
}
\description{
Use the limma or DESeq2 package to perform DEG analysis
with the specified model design. Core function for the
DEG panel of iDEP.
}
