% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{list_factors_ui}
\alias{list_factors_ui}
\title{Create choices and title for model factors}
\usage{
list_factors_ui(sample_info, data_file_format, counts_deg_method)
}
\arguments{
\item{sample_info}{Experiment file information for grouping}

\item{data_file_format}{Type of data being examined}

\item{counts_deg_method}{Method of DEG being performed (See
DEG UI for options)}
}
\value{
A list containing a string title and a vector of
 comparisons to choose from.
}
\description{
Create a list of options for comparisons from the 
sample_info. Also create a title to be used in a 
UI checkbox based off the date file format and the
DEG method. The list of options will become
checkboxes for the user to create their experiment
design from.
}
