% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{list_interaction_terms_ui}
\alias{list_interaction_terms_ui}
\title{List model interaction terms}
\usage{
list_interaction_terms_ui(sample_info, select_factors_model)
}
\arguments{
\item{sample_info}{Experiment file information for grouping}

\item{select_factors_model}{The selected factors for the model
expression}
}
\value{
Returns a character string of an interaction term
 between the selected factors. Used in a checkbox for the
 User to create a model expression
}
\description{
This functions uses the sample info file and the selected
model factors to create interaction terms to be used in the 
DEG process.
}
