% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{make_pull_down_menu}
\alias{make_pull_down_menu}
\title{make_pull_down_menu makes it easy to make pull down menu}
\usage{
make_pull_down_menu(funs)
}
\arguments{
\item{funs}{}
}
\value{

}
\description{

}
