% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{mean_sd_plot}
\alias{mean_sd_plot}
\title{Mean vs. Standard Deviation plot}
\usage{
mean_sd_plot(processed_data, rank, heat_cols)
}
\arguments{
\item{processed_data}{Data that has gone through the pre-processing}

\item{rank}{TRUE/FALSE whether to use the rank of the mean or not}

\item{heat_cols}{Heat color to use with black in the plot}
}
\value{
A formatted ggplot hexplot of the mean and standard
 deviation of the processed data
}
\description{
Create a plot that shows the standard deviation as the
Y-axis across the mean of the counts data on the X-axis.
Option to make the X-axis the rank of the mean which
does a better job showing the spread of the data.
}
