% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{merge_data}
\alias{merge_data}
\title{Merge data from gene info and processed data}
\usage{
merge_data(all_gene_names, data, merge_ID)
}
\arguments{
\item{all_gene_names}{All matched gene names from idep data}

\item{data}{Data matrix with rownames to merge with gene names}
}
\value{
Inputted data with all gene name information.
}
\description{
This function takes in the gene info data and merges it
with the data that has gone through the processing fcn.
The returned data contains the gene names as well as the
ensembl id in the first two columns.
}
