% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{network_data}
\alias{network_data}
\title{VisNetwork data}
\usage{
network_data(
  network,
  up_down_reg_deg,
  wrap_text_network_deg,
  layout_vis_deg,
  edge_cutoff_deg
)
}
\arguments{
\item{network}{GO table from the pathway analysis}

\item{up_down_reg_deg}{Plot just up/down or both}

\item{wrap_text_network_deg}{Wrap the text from the pathway description}

\item{layout_vis_deg}{BUtton to reset the layout of the network}

\item{edge_cutoff_deg}{P-value to cutoff enriched pathways}
}
\value{
Data that can be inputted in the vis_network_plot function
 to create an interactive network.
}
\description{
Create VisNetwork data that can be inputted in the vis_network_plot
function to create an interactive network of enriched pathways.
}
