% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_09_network.R
\name{network_enrich_data}
\alias{network_enrich_data}
\title{Gene vector query for enrichment}
\usage{
network_enrich_data(select_wgcna_module, wgcna)
}
\arguments{
\item{wgcna}{List returned from the \code{get_wgcna}}

\item{select_wgcna_modules}{The module to create a plot of hte top genes for,
options can be found with the \code{get_wgcna_modules} function}
}
\value{
A vector of genes that are included in the selected module.
}
\description{
Select a module to create a vector of gene IDs to use in an enrichment
analysis.
}
