% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{pathway_select_data}
\alias{pathway_select_data}
\title{Get data from genes in selected pathway}
\usage{
pathway_select_data(
  sig_pathways,
  gene_sets,
  contrast_samples,
  data,
  select_org,
  all_gene_names
)
}
\arguments{
\item{sig_pathways}{Description of the pathway for which to
obtain the gene expression data}

\item{gene_sets}{List of vectors with each vector being the
set of genes that correspond to a particular pathway in
the database (from read_gene_sets function)}

\item{contrast_samples}{Sample columns that correspond to the
selected comparison}

\item{data}{Data that has been through the pre-processing}

\item{select_org}{The organism that the gene data is for}

\item{all_gene_names}{Matrix of all the matched and converted
gene IDs}
}
\value{
Sub-data matrix from the processed data. Only contains
 genes from the selected pathway and samples that correspond to
 the comparison being analyzed.
}
\description{
Return a data matrix that is a subset of the processed data and
only contains genes that are in the gene set of the desired
pathway.
}
