% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{pc_factor_correlation}
\alias{pc_factor_correlation}
\title{Correlations Between Principle Components and Factors}
\usage{
pc_factor_correlation(data, sample_info)
}
\arguments{
\item{data}{Data that has been through pre-processing}

\item{sample_info}{Matrix array with experiment info}
}
\value{
text with correlation
}
\description{
Desc
}
