% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{PCA_biplot}
\alias{PCA_biplot}
\title{Principal Component Analysis with PCAtools package}
\usage{
PCA_biplot(
  data,
  sample_info,
  select_gene_id = "symbol",
  all_gene_names,
  selected_x = "PC1",
  selected_y = "PC2",
  encircle = TRUE,
  encircleFill = TRUE,
  showLoadings = TRUE,
  pointlabs = TRUE,
  point_size = 4,
  ui_color = NULL,
  ui_shape = NULL
)
}
\arguments{
\item{data}{Data that has been through pre-processing}

\item{sample_info}{Matrix array with experiment info}
}
\value{
Formatted PCA plot using PCAtools package
}
\description{
Draw a PCA plot using PCAtools package
}
