% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{PCA_plot}
\alias{PCA_plot}
\title{Principal Component Analysis}
\usage{
PCA_plot(
  data,
  sample_info,
  PCAx = 1,
  PCAy = 2,
  selected_color = "Sample_Name",
  selected_shape = "Sample_Name"
)
}
\arguments{
\item{data}{Data that has been through pre-processing}

\item{sample_info}{Matrix array with experiment info}

\item{PCAx}{PC on x axis}

\item{PCAy}{PC on y axis}
}
\value{
Formatted PCA plot
}
\description{
Draw a PCA plot where user selects which PCs on axes
}
