% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{PCAtools_eigencorplot}
\alias{PCAtools_eigencorplot}
\title{Principal Component Analysis with PCAtools package}
\usage{
PCAtools_eigencorplot(processed_data, sample_info)
}
\arguments{
\item{sample_info}{Design Matrix}

\item{data}{Data that has been through pre-processing}
}
\value{
Formatted plot generated with PCAtools package
}
\description{
Generates a plot showing correlations between Principal Components and design factors
}
