% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_06_pathway.R
\name{pgsea_plot_all}
\alias{pgsea_plot_all}
\title{Pathway analysis with the PGSEA package on all samples}
\usage{
pgsea_plot_all(
  go,
  my_range,
  data,
  select_contrast,
  gene_sets,
  pathway_p_val_cutoff,
  n_pathway_show
)
}
\arguments{
\item{go}{Portion of the database to use for the pathway analysis}

\item{my_range}{Vector of the (min_set_size, max_set_size)}

\item{data}{Data that has been through the pre-processing}

\item{select_contrast}{Comparison from DEG analysis to
use the top genes from in the pathway analysis}

\item{gene_sets}{List of vectors with each vector being the
set of genes that correspond to a particular pathway in
the database (from read_gene_sets function)}

\item{pathway_p_val_cutoff}{Significant p-value to determine
enriched pathways}

\item{n_pathway_show}{Number of pathways to return in final
result}
}
\value{
A data frame with the results of the pathway analysis.
 The data frame has five columns for the direction of the
 regulation, the pathway description, the stat value, the
 number of overlapping genes, and the p-value.
}
\description{
Run pathway analysis with the PGSEA package using the results
from the limma_value function on all samples.
}
