% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{plot_deg_scatter}
\alias{plot_deg_scatter}
\title{Scatter plot of the comparison groups}
\usage{
plot_deg_scatter(
  select_contrast,
  comparisons,
  top_genes,
  limma_p_val,
  limma_fc,
  contrast_samples,
  processed_data,
  sample_info
)
}
\arguments{
\item{select_contrast}{Comparison from DEG analysis to filter
for the significant genes}

\item{comparisons}{The comparisons vector from the results list
of the limma_value function}

\item{top_genes}{top_genes list from results list of the
limma_value function}

\item{limma_p_val}{Significant p-value to use to in determining
the expressed genes}

\item{limma_fc}{Minimum fold change value to use in determining
the expressed genes}

\item{contrast_samples}{Samples that are included in the selected
comparison}

\item{processed_data}{Data matrix that has gone through
pre-processing}

\item{sample_info}{Experiment file information for grouping}
}
\value{
A formatted ggplot with the X-axis as the mean expression
 of one contrast group and the Y-axis as the mean expression of
 the other contrast group.
}
\description{
Create a scatter plot of the expression value for each gene 
in the two groups for the selected contrast. For the selected
contrast, the mean expression is calculated for a gene in both
group of samples in the contrast and plotted in a scatter plot.
}
