% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{plot_ma}
\alias{plot_ma}
\title{Plot mean expression and fold change}
\usage{
plot_ma(
  select_contrast,
  comparisons,
  top_genes,
  limma_p_val,
  limma_fc,
  contrast_samples,
  processed_data,
  plot_colors
)
}
\arguments{
\item{select_contrast}{Comparison from DEG analysis to filter
for the significant genes}

\item{comparisons}{The comparisons vector from the results list
of the limma_value function}

\item{top_genes}{top_genes list from results list of the
limma_value function}

\item{limma_p_val}{Significant p-value to use to in determining
the expressed genes}

\item{limma_fc}{Minimum fold change value to use in determining
the expressed genes}

\item{contrast_samples}{Samples that are included in the selected
comparison}

\item{processed_data}{Data matrix that has gone through
pre-processing}

\item{plot_colors}{List containing three colors to differentiate between
the up-regulated, down-regulated, and other genes}
}
\value{
A ggplot with the X-axis the mean expression value and
 the Y-axis the calculated fold-change from the DEG analysis.
}
\description{
Draw a ggplot of the overal mean expression for each gene 
and the calculated fold change for the selected comparison.
}
