% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_09_network.R
\name{plot_mean_connectivity}
\alias{plot_mean_connectivity}
\title{Mean connectivity plot}
\usage{
plot_mean_connectivity(wgcna)
}
\arguments{
\item{wgcna}{List returned from the \code{get_wgcna}}
}
\value{
A formatted ggplot displaying the mean connectivityfor the
 \code{get_wgcna} function return.
}
\description{
Create a ggplot from the wgcna object to display the mean connectivity.
}
