% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_09_network.R
\name{plot_scale_independence}
\alias{plot_scale_independence}
\title{Scale independence plot}
\usage{
plot_scale_independence(wgcna)
}
\arguments{
\item{wgcna}{List returned from the \code{get_wgcna}}
}
\value{
A formatted ggplot displaying the scale independence for the
 \code{get_wgcna} function return.
}
\description{
Using the WGCNA return, create a ggplot of scale independence.
}
