% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{pre_process}
\alias{pre_process}
\title{Pre-Process the data}
\usage{
pre_process(
  data,
  missing_value,
  data_file_format,
  low_filter_fpkm,
  n_min_samples_fpkm,
  log_transform_fpkm,
  log_start_fpkm,
  min_counts,
  n_min_samples_count,
  counts_transform,
  counts_log_start,
  no_fdr
)
}
\arguments{
\item{data}{Data that has already gone through the convert_data fcn}

\item{missing_value}{Method to deal with missing data}

\item{data_file_format}{Type of data being examined}

\item{low_filter_fpkm}{Low count filter for the fpkm data}

\item{n_min_samples_fpkm}{Min samples for fpkm data}

\item{log_transform_fpkm}{Type of transformation for fpkm data}

\item{log_start_fpkm}{Value added to log transformation for fpkm}

\item{min_counts}{Low count filter for count data}

\item{n_min_samples_count}{Min sample for count data}

\item{counts_transform}{Type of transformation for counts data}

\item{counts_log_start}{Value added to log for counts data}

\item{no_fdr}{Fold changes only data with no p values}
}
\value{
A list containing the transformed data, the mean kurtosis,
the raw counts, a data type warning, the size of the original data,
and p-values.
}
\description{
This function takes in user defined values to
process the data for the EDA that occurs on the second page.
All the filtering and transformation that the app does will
occur on this page.
}
