% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{proper}
\alias{proper}
\title{FUNCTION_TITLE}
\usage{
proper(x)
}
\arguments{
\item{x}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
