% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{read_gene_sets}
\alias{read_gene_sets}
\title{Read pathway gene sets}
\usage{
read_gene_sets(converted, all_gene_names, go, select_org, idep_data, my_range)
}
\arguments{
\item{converted}{Return value from the \code{convert_id} function,
contains information about the gene IDs for the matched species}

\item{all_gene_names}{Data frame with all gene names}

\item{go}{Portion of the database to use for the pathway analysis}

\item{select_org}{The user selected organism for the expression data,
default is "BestMatch"}

\item{idep_data}{Read data files from the database}

\item{my_range}{Vector of the (min_set_size, max_set_size)}
}
\value{
A list with each entry a list of gene IDs that correspond to
 a pathway.
}
\description{
Use the IDs from the converted database to find the
gene sets for all the pathways in the database. Returns
a list with each entry a vector of IDs corresponding to
a description of a pathway in the database.
}
