% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{read_gmt}
\alias{read_gmt}
\title{FUNCTION_TITLE}
\usage{
read_gmt(file_path)
}
\arguments{
\item{file_path}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
