% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{read_pathway_sets}
\alias{read_pathway_sets}
\title{Read pathway sets for gene query}
\usage{
read_pathway_sets(
  all_gene_names_query,
  converted,
  go,
  select_org,
  gmt_file,
  idep_data,
  gene_info
)
}
\arguments{
\item{all_gene_names_query}{Subsetted rows of the 
all_gene_names data frame to query the database with}

\item{converted}{Conversion information from the original
IDs returned from the convert_id() function}

\item{go}{Section of the database to query for pathway
analysis}

\item{select_org}{Input for what organism the IDs are 
pertaining to}

\item{gmt_file}{For NEW species the gmt file to use
for the pathway analysis}

\item{idep_data}{Data built in to idep}

\item{gene_info}{The gene info from the converted IDs and
the function gene_info()}
}
\value{
This function returns a list with values that are
  used in the find_overlap function. The list contains
  pathway_table which is the overlap and total genes for
  each pathway that is enriched in the query. The list
  also contains the query_set of genes, the total_genes
  number which is used in the calculation of the p-values
  and the pathway files that contain gmt information on
  the mathced species.
}
\description{
This function provides the gene set information
to perform enrichment analysis on the provided
query.
}
