% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis_random.R
\name{render.kegg.node}
\alias{render.kegg.node}
\title{PATHVIEW SOURCE FUNCTION
https://rdrr.io/bioc/pathview/src/R/render.kegg.node.R}
\usage{
render.kegg.node(
  plot.data,
  cols.ts,
  img,
  same.layer = TRUE,
  type = c("gene", "compound")[1],
  text.col = "black",
  cex = 0.25
)
}
\description{
PATHVIEW SOURCE FUNCTION
https://rdrr.io/bioc/pathview/src/R/render.kegg.node.R
}
