% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_03_clustering.R
\name{sd_density}
\alias{sd_density}
\title{Density plot of data standard deviation}
\usage{
sd_density(data, n_genes_max)
}
\arguments{
\item{data}{Data that has been through pre-processing}

\item{n_genes_max}{Upper limit of gene range}
}
\value{
Formatted density plot of the standard deviation
distribution.
}
\description{
Draw a denisty plot of the standard deviation in the
data. Add vertical red lines for a range of genes.
}
