% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{select_reference_levels_ui}
\alias{select_reference_levels_ui}
\title{Refernce levels for selected factor}
\usage{
select_reference_levels_ui(
  sample_info,
  select_factors_model,
  data_file_format,
  counts_deg_method
)
}
\arguments{
\item{sample_info}{Experiment file information for grouping}

\item{select_factors_model}{The selected factors for the model
expression}

\item{data_file_format}{Type of gene data being examined}

\item{counts_deg_method}{The method or package being used for
the DEG analysis}
}
\value{
A list the same length as the vector of selected factors.
 Each entry in the list corresponds to the choice of group to
 use for the reference level.
}
\description{
This function uses a vector of selected factors to create
choices for the reference level to use for the factor in
the DEG analysis.
}
