% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{sig_genes_plot}
\alias{sig_genes_plot}
\title{Significant genes bar plot}
\usage{
sig_genes_plot(results)
}
\arguments{
\item{results}{Results matrix from the limma_value function
returned list}
}
\value{
Formatted gg barplot of the significantly expressed
 genes.
}
\description{
Create a bar plot of the number of genes with a significant fold
change. The plot will break down all of the comparisons that
were analyzed and the count of significant genes for both up
and down changes.
}
