% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_05_deg.R
\name{split_interaction_terms}
\alias{split_interaction_terms}
\title{SPLIT INTERACION TERMS
Split  genotypeI.conditionTrt --> c("genotype","I","conditoin","Trt")}
\usage{
split_interaction_terms(term, factor_lookup, level_lookup)
}
\description{
SPLIT INTERACION TERMS
Split  genotypeI.conditionTrt --> c("genotype","I","conditoin","Trt")
}
