% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_04_pca.R
\name{t_SNE_plot}
\alias{t_SNE_plot}
\title{TSNE FUNCTION}
\usage{
t_SNE_plot(data, sample_info, selected_color, selected_shape)
}
\arguments{
\item{data}{Data that has been through pre-processing}

\item{sample_info}{Matrix array with experiment info}
}
\value{
Formatted T-sne plot
}
\description{
Draw a t-sne plot where user selects which PCs on axes
}
