% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_02_pre_process.R
\name{total_counts_ggplot}
\alias{total_counts_ggplot}
\title{Creates a barplot of the count data}
\usage{
total_counts_ggplot(counts_data, sample_info)
}
\arguments{
\item{counts_data}{Raw counts from gene expression data}

\item{sample_info}{Experiment file information for grouping
samples}
}
\value{
formatted ggbarplot
}
\description{
This function takes in the rount count data
and creates a formatted gg barplot that shows the number
of genes mapped to each sample in millions.
}
