% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{utils_ui.R}
\alias{utils_ui.R}
\title{utils_ui.R These are helper functions for ui code,
shiny code can go in this file}
\description{
utils_ui.R These are helper functions for ui code,
shiny code can go in this file
}
\section{utils_ui.R functions}{

\code{make_pull_down_menu} makes it easy to make pull down menu
}

